requirements.txt
flask>=3.0.0
PyPDF2>=3.0.0
groq>=0.4.0
werkzeug>=3.0.0
requests>=2.31.0
supabase>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
gunicorn>=21.0.0
Pillow>=10.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
python-dotenv>=1.0.1
